feat(home): home feature moved
