+++ apps/startup-glossary/src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 const routes: Routes = [
     {
         path: 'home',
-        loadChildren: () => import('./home/home.module').then(m => m.HomePageModule),
+        loadChildren: () => import('@mobile-academy/features/glossary-home-module').then(m => m.GlossaryHomeModule),
     },
     {
         path: 'category',
