--- libs/features/glossary-home-module/src/lib/features-glossary-home-module.spec.ts
@@ -1,7 +1,0 @@
-import { featuresGlossaryHomeModule } from './features-glossary-home-module';
-
-describe('featuresGlossaryHomeModule', () => {
-    it('should work', () => {
-        expect(featuresGlossaryHomeModule()).toEqual('features-glossary-home-module');
-    });
-});
