--- libs/features/glossary-home-module/src/lib/features-glossary-home-module.ts
@@ -1,3 +1,0 @@
-export function featuresGlossaryHomeModule(): string {
-    return 'features-glossary-home-module';
-}
