+++ libs/features/glossary-home-module/src/lib/glossary-home-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomePage } from './home.page';
+import { HomePage } from './pages/root/home.page';
 
 const routes: Routes = [
     {
@@ -13,4 +13,4 @@ const routes: Routes = [
     imports: [RouterModule.forChild(routes)],
     exports: [RouterModule],
 })
-export class HomePageRoutingModule {}
+export class GlossaryHomeRoutingModule {}
