+++ libs/features/glossary-home-module/src/lib/glossary-home.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsModule } from '@angular/forms';
-import { HomePage } from './home.page';
+import { HomePage } from './pages/root/home.page';
 import { CategoryService } from '@mobile-academy/data-access/services';
-import { HomePageRoutingModule } from './home-routing.module';
+import { GlossaryHomeRoutingModule } from './glossary-home-routing.module';
 
 @NgModule({
-    imports: [CommonModule, FormsModule, IonicModule, HomePageRoutingModule],
+    imports: [CommonModule, FormsModule, IonicModule, GlossaryHomeRoutingModule],
     declarations: [HomePage],
     providers: [CategoryService],
 })
-export class HomePageModule {}
+export class GlossaryHomeModule {}
