+++ libs/features/glossary-home-module/src/lib/pages/root/home.page.scss
