+++ tsconfig.base.json
@@ -19,7 +19,7 @@
             "@mobile-academy/data-access/models": ["libs/data-access/models/src/index.ts"],
             "@mobile-academy/data-access/services": ["libs/data-access/services/src/index.ts"],
             "@mobile-academy/features/glossary-category-module": ["libs/features/glossary-category-module/src/index.ts"],
-            "@mobile-academy/features/glossary-home-module": ["libs/features/glossary-home-module/src/index.ts"],
+            "@mobile-academy/features/glossary-home-module": ["libs/features/glossary-home-module/src/lib/glossary-home.module"],
             "@mobile-academy/ui/components": ["libs/ui/components/src/index.ts"],
             "@mobile-academy/util/enums": ["libs/util/enums/src/index.ts"]
         }
