+++ tsconfig.json
@@ -1,26 +1,28 @@
 {
-    "compileOnSave": false,
-    "compilerOptions": {
-        "rootDir": ".",
-        "sourceMap": true,
-        "declaration": false,
-        "moduleResolution": "node",
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true,
-        "importHelpers": true,
-        "target": "es2015",
-        "module": "esnext",
-        "lib": ["es2017", "dom"],
-        "skipLibCheck": true,
-        "skipDefaultLibCheck": true,
-        "baseUrl": ".",
-        "paths": {
-            "@mobile-academy/data-access/glossary": ["libs/data-access/glossary/src/index.ts"],
-            "@mobile-academy/data-access/models": ["libs/data-access/models/src/index.ts"],
-            "@mobile-academy/data-access/services": ["libs/data-access/services/src/index.ts"],
-            "@mobile-academy/ui/components": ["libs/ui/components/src/index.ts"],
-            "@mobile-academy/util/enums": ["libs/util/enums/src/index.ts"]
-        }
-    },
-    "exclude": ["node_modules", "tmp"]
+  "compileOnSave": false,
+  "compilerOptions": {
+      "rootDir": ".",
+      "sourceMap": true,
+      "declaration": false,
+      "moduleResolution": "node",
+      "emitDecoratorMetadata": true,
+      "experimentalDecorators": true,
+      "importHelpers": true,
+      "target": "es2015",
+      "module": "esnext",
+      "lib": ["es2017", "dom"],
+      "skipLibCheck": true,
+      "skipDefaultLibCheck": true,
+      "baseUrl": ".",
+      "paths": {
+          "@mobile-academy/data-access/glossary": ["libs/data-access/glossary/src/index.ts"],
+          "@mobile-academy/data-access/models": ["libs/data-access/models/src/index.ts"],
+          "@mobile-academy/data-access/services": ["libs/data-access/services/src/index.ts"],
+          "@mobile-academy/features/glossary-category-module": ["libs/features/glossary-category-module/src/index.ts"],
+          "@mobile-academy/features/glossary-home-module": ["libs/features/glossary-home-module/src/lib/glossary-home.module"],
+          "@mobile-academy/ui/components": ["libs/ui/components/src/index.ts"],
+          "@mobile-academy/util/enums": ["libs/util/enums/src/index.ts"]
+      }
+  },
+  "exclude": ["node_modules", "tmp"]
 }
